pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fabric-mcp-server-powerbi-creator"
version = "0.1.0"
description = "MCP server for Microsoft Fabric workspace management, Git-based deployments, and Power BI theme generation"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.25.0",
    "requests",
    "pydantic",
    "azure-identity",
    "azure-storage-file-datalake"
]

[project.scripts]
fabric-mcp-server-powerbi-creator = "fabric_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/fabric_mcp"]
